** !! BREAK CHANGE !!  ** MISC : relocate some util classes
